reward_server/pyproject.toml
[project]
name = "service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.124.2",
    "hpsv2>=1.2.0",
    "image-reward>=1.5",
    "numpy>=2.2.6",
    "open-clip-torch>=2.32.0",
    "pillow>=12.0.0",
    "python-multipart>=0.0.20",
    "torch>=2.1.0",
    "urllib3>=2.6.1",
    "uvicorn>=0.38.0",
]

